+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- **Added:** A new Sound Manager tab for playing around with sounds.
+- **Added:** Highlighting a positional SoundData entry will draw a line to its position.
 - **Updated:** ClientStructs now at [008914bf](https://github.com/aers/FFXIVClientStructs/tree/008914bf) ([compare](https://github.com/aers/FFXIVClientStructs/compare/302aacc9..008914bf)).
 
 ## [1.49.0] (2026-02-14)